feat(website): redeem page
