+++ apps/website/src/routes/redeem/+page.svelte
@@ -1,0 +1,1 @@
+Redeem
